feat: kickstart plugins and custom plugins
