+++ .gitignore
@@ -3,5 +3,4 @@ test.sh
 .luarc.json
 nvim
 
-spell/
 lazy-lock.json
