+++ lua/custom/plugins/hop.lua
@@ -1,0 +1,9 @@
+return {
+  {
+    'phaazon/hop.nvim',
+    branch = 'v2',
+    config = function()
+      require('hop').setup()
+    end,
+  },
+}
