+++ lua/custom/plugins/rust-crates.lua
@@ -1,0 +1,18 @@
+return {
+  {
+    'saecki/crates.nvim',
+    ft = { 'toml' },
+    config = function()
+      require('crates').setup {
+        completion = {
+          cmp = {
+            enabled = true,
+          },
+        },
+      }
+      require('cmp').setup.buffer {
+        sources = { { name = 'crates' } },
+      }
+    end,
+  },
+}
