+++ lua/custom/plugins/rust.lua
@@ -1,0 +1,9 @@
+return {
+  {
+    'rust-lang/rust.vim',
+    ft = 'rust',
+    init = function()
+      vim.g.rustfmt_autosave = 0
+    end,
+  },
+}
