+++ lua/kickstart/plugins/neo-tree.lua
@@ -20,6 +20,9 @@ return {
           ['\\'] = 'close_window',
         },
       },
+      filtered_items = {
+        hide_dotfiles = false,
+      },
     },
   },
 }
