+++ lua/kickstart/plugins/todo-comments.lua
@@ -1,0 +1,3 @@
+return {
+  { 'folke/todo-comments.nvim', event = 'VimEnter', dependencies = { 'nvim-lua/plenary.nvim' }, opts = { signs = false } },
+}
