+++ lua/kickstart/plugins/vim-sleuth.lua
@@ -1,0 +1,5 @@
+return {
+  {
+    'tpope/vim-sleuth',
+  }
+}
