+++ spell/en.utf-8.add
@@ -1,0 +1,217 @@
+RABBITMQ
+POSTGRES
+postgres
+localhost
+hdr
+AMQP
+nvim
+Makefile
+README
+json
+vscode
+github
+deps
+md
+github
+QueueDefinition
+cfgs
+dlq
+AmqpTopology
+rmq
+ExchangeDefinition
+acks
+QueueBinding
+Mastercard
+println/!
+mqtt
+jsonwebtoken
+moka
+neotest
+noice
+rustaceanvim
+sanegx
+chadrc
+stylua
+lspconfig
+ftplugin
+codelldb
+liblldb
+mrcjkb
+davidmh
+nvimtools
+hedro
+psql
+proteu
+gerdau
+ONPREMISSES
+onpremise
+FARGATE
+awsvpc
+hedrogcp
+awslogs
+tdef
+awscredentials
+localstack
+noremap
+bufnr
+dapui
+Lazygit
+startinsert
+termopen
+gopls
+goimports
+golines
+tohtml
+getscript
+getscript
+logipat
+netrw
+netrw
+matchit
+rrhelper
+spellfile
+vimball
+rplugin
+synmenu
+optwin
+bugreport
+nvlsp
+nvchad
+terraformls
+cssls
+gofumpt
+codespell
+Gitsigns
+keymap
+grpcs
+nvimtree
+tecnology
+tfvars
+appuser
+asapcard
+manufacurerkey
+manufacurer
+moov
+paymentcryptographydata
+paymentcryptography
+asapshared
+logrus
+Ecommerce
+entrepay
+Tdes
+Manufactores
+configtool
+amqps
+sirupsen
+goopentelemetrylib
+otellogrus
+gomultisqldbpg
+Warnf
+Infof
+mgmt
+discoverytool
+discoverymgmt
+Debugf
+rdkafka
+rdkafka
+traceparent
+tracestate
+relativenumber
+DMC
+ELO
+TRX
+Enum
+rabbitmq
+CPNJ
+config
+multisqldb
+adempiere
+mastercard
+timestamp
+trxRepository
+trx
+redis
+elo
+vsplit
+startify
+rssi
+accraw
+Prtv
+ParseRMMSError
+rmms
+mockall
+truststore
+sasl
+CNPJ
+clientKey
+protos
+viewmodels
+actix_web
+utoipa
+deadpool_postgres
+chrono
+hedrosistemas
+hedrosistemas
+tlsv1
+Toolchain
+rustup
+allowedby
+foldmethod
+foldexpr
+foldenable
+nvim_create_autocmd
+texthl
+linehl
+numhl
+Nvdash
+spelllang
+thiserror
+smoka7
+multicursors
+guifg
+guibg
+Auth
+stevearc
+vertsplit
+nord_blue
+statusline_bg
+lightbg
+pmenu_bg
+bg
+treesiter
+chadracula
+pyckeycleanup
+auth
+lazydocker
+CybersourceSDK
+cybersource_sdk
+cybersourceHost
+cybersourceMerchantId
+cybersourceBaseSDK
+cybersource
+devicons
+MunifTanjim
+avante
+yetone
+fanout
+GatewayCiotStatus
+GatewayUartStatus
+ciot_status
+uart_status
+uart
+ciot
+gomod
+gofmt
+clippy
+taplo
+dockerfmt
+struct
+vimdoc
+gosum
+gowork
+nvim
+nvim
+txt
+folke
+https
+gitsigns
